Attempted to fix EFI_NOT_FOUND on kernel load

+++ kernel/core/include/memory.h
@@ -7,7 +7,7 @@ typedef uint64_t physical_address_t;
 typedef uint64_t virtual_address_t;
 
 typedef struct {
-    uint64_t                      Type;           // Field size is 32 bits followed by 32 bit pad
+    uint32_t                      Type;           // Field size is 32 bits followed by 32 bit pad
     uint64_t                      Pad;
     physical_address_t            PhysicalStart;  // Field size is 64 bits
     virtual_address_t             VirtualStart;   // Field size is 64 bits
